Specialize eval-primo and provide missing booleans
The possibilities for #t and #f values were missing in 'pair?' and
'symbol?' handling.

In the specialization test this achieves a 5x speedup on the 'equal?'
benchmarks, but not much changes in the normal setting.  There seems to
be a bad interaction with var-args procedures: when they're
re-introduced in the specialization test, the performance gain
disappears.